fix: removed DefaultReporterUpgrader
fixes AM-7461
ref gravitee-io/issues#11716
