+++ gravitee-am-gateway/gravitee-am-gateway-handler/gravitee-am-gateway-handler-common/src/main/java/io/gravitee/am/gateway/handler/common/audit/impl/GatewayAuditReporterManager.java
@@ -31,6 +31,7 @@ import io.gravitee.am.plugins.reporter.core.ReporterPluginManager;
 import io.gravitee.am.plugins.reporter.core.ReporterProviderConfiguration;
 import io.gravitee.am.repository.management.api.ReporterRepository;
 import io.gravitee.am.service.EnvironmentService;
+import io.gravitee.am.service.reporter.SystemReporterConfigResolver;
 import io.gravitee.am.service.reporter.impl.AuditReporterVerticle;
 import io.gravitee.am.service.reporter.vertx.EventBusReporterWrapper;
 import io.gravitee.common.event.Event;
@@ -75,6 +76,9 @@ public class GatewayAuditReporterManager extends AbstractService<AuditReporterMa
 
     @Autowired
     private EnvironmentService environmentService;
+
+    @Autowired
+    private SystemReporterConfigResolver systemReporterConfigResolver;
 
     private final ConcurrentMap<String, io.gravitee.am.reporter.api.provider.Reporter> reporterPlugins = new ConcurrentHashMap<>();
     private final ConcurrentMap<String, Reporter> reporters = new ConcurrentHashMap<>();
@@ -235,7 +239,14 @@ public class GatewayAuditReporterManager extends AbstractService<AuditReporterMa
             return false;
         }
         logger.info("\tInitializing reporter: {} [{}]", reporter.getName(), reporter.getType());
-        var providerConfiguration = new ReporterProviderConfiguration(reporter, context);
+        Reporter resolved;
+        try {
+            resolved = systemReporterConfigResolver.resolve(reporter);
+        } catch (Exception ex) {
+            logger.error("Unable to resolve configuration for reporter {}", reporter.getId(), ex);
+            return false;
+        }
+        var providerConfiguration = new ReporterProviderConfiguration(resolved, context);
         io.gravitee.am.reporter.api.provider.Reporter reporterProvider = reporterPluginManager.create(providerConfiguration);
 
         if (reporterProvider != null) {
