+++ gravitee-am-management-api/gravitee-am-management-api-service/src/main/java/io/gravitee/am/management/service/impl/ReporterServiceProxyImpl.java
@@ -111,11 +111,6 @@ public class ReporterServiceProxyImpl extends AbstractSensitiveProxy implements 
     }
 
     @Override
-    public String createReporterConfig(Reference reference){
-        return reporterService.createReporterConfig(reference);
-    }
-
-    @Override
     public Completable notifyInheritedReporters(Reference parentReference, Reference affectedReference, Action action) {
         return reporterService.notifyInheritedReporters(parentReference, affectedReference, action);
     }
