+++ gravitee-am-service/src/main/java/io/gravitee/am/service/impl/ReporterServiceImpl.java
@@ -16,20 +16,16 @@
 package io.gravitee.am.service.impl;
 
 import com.google.common.base.Strings;
-import com.google.common.io.BaseEncoding;
 import io.gravitee.am.common.audit.EventType;
-import io.gravitee.am.common.env.RepositoriesEnvironment;
 import io.gravitee.am.common.event.Action;
 import io.gravitee.am.common.event.Type;
 import io.gravitee.am.common.utils.RandomString;
 import io.gravitee.am.identityprovider.api.User;
-import io.gravitee.am.model.Organization;
 import io.gravitee.am.model.Reference;
 import io.gravitee.am.model.ReferenceType;
 import io.gravitee.am.model.Reporter;
 import io.gravitee.am.model.common.event.Event;
 import io.gravitee.am.model.common.event.Payload;
-import io.gravitee.am.repository.Scope;
 import io.gravitee.am.repository.management.api.ReporterRepository;
 import io.gravitee.am.service.AuditService;
 import io.gravitee.am.service.EventService;
@@ -42,6 +38,7 @@ import io.gravitee.am.service.exception.ReporterNotFoundException;
 import io.gravitee.am.service.exception.TechnicalManagementException;
 import io.gravitee.am.service.model.NewReporter;
 import io.gravitee.am.service.model.UpdateReporter;
+import io.gravitee.am.service.reporter.SystemReporterConfigResolver;
 import io.gravitee.am.service.reporter.builder.AuditBuilder;
 import io.gravitee.am.service.reporter.builder.management.ReporterAuditBuilder;
 import io.reactivex.rxjava3.core.Completable;
@@ -55,19 +52,10 @@ import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
 
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-import java.util.ArrayList;
 import java.util.Date;
-import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-
-import static io.gravitee.am.service.utils.BackendConfigurationUtils.getMongoDatabaseName;
 
 /**
  * @author Titouan COMPIEGNE (titouan.compiegne at graviteesource.com)
@@ -78,17 +66,13 @@ import static io.gravitee.am.service.utils.BackendConfigurationUtils.getMongoDat
 public class ReporterServiceImpl implements ReporterService {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ReporterServiceImpl.class);
-    private static final int TABLE_SUFFIX_MAX_LENGTH = 30;
-    private static final String REPORTER_AM_JDBC = "reporter-am-jdbc";
     private static final String REPORTER_AM_FILE = "reporter-am-file";
     private static final String REPORTER_CONFIG_FILENAME = "filename";
-    public static final String MANAGEMENT_TYPE = Scope.MANAGEMENT.getRepositoryPropertyKey() + ".type";
-    public static final String MONGODB = "mongodb";
     // Regex as defined into the Reporter plugin schema in order to apply the same validation rule
     // when a REST call is performed and not only check on the UI
     private final Pattern filenamePattern = Pattern.compile("^([A-Za-z0-9][A-Za-z0-9\\-_.]*)$");
 
-    private RepositoriesEnvironment environment;
+    private SystemReporterConfigResolver systemReporterConfigResolver;
 
     private ReporterRepository reporterRepository;
 
@@ -98,8 +82,8 @@ public class ReporterServiceImpl implements ReporterService {
 
     private PluginConfigurationValidationService validationService;
 
-    public ReporterServiceImpl(RepositoriesEnvironment environment, @Lazy ReporterRepository reporterRepository, EventService eventService, AuditService auditService, PluginConfigurationValidationService validationService) {
-        this.environment = environment;
+    public ReporterServiceImpl(SystemReporterConfigResolver systemReporterConfigResolver, @Lazy ReporterRepository reporterRepository, EventService eventService, AuditService auditService, PluginConfigurationValidationService validationService) {
+        this.systemReporterConfigResolver = systemReporterConfigResolver;
         this.reporterRepository = reporterRepository;
         this.eventService = eventService;
         this.auditService = auditService;
@@ -142,20 +126,14 @@ public class ReporterServiceImpl implements ReporterService {
         LOGGER.debug("Create default reporter for  {}", reference);
         NewReporter newReporter = createInternal(reference);
         if (newReporter == null) {
-            return Single.error(new ReporterNotFoundException("Reporter type " + this.environment.getProperty(MANAGEMENT_TYPE) + " not found"));
+            return Single.error(new ReporterNotFoundException("Reporter type " + systemReporterConfigResolver.managementBackend() + " not found"));
         }
         return create(reference, newReporter);
     }
 
     @Override
     public NewReporter createInternal(Reference reference) {
-        NewReporter newReporter = null;
-        if (useMongoReporter()) {
-            newReporter = createMongoReporter(reference);
-        } else if (useJdbcReporter()) {
-            newReporter = createJdbcReporter(reference);
-        }
-        return newReporter;
+        return systemReporterConfigResolver.createInternal(reference);
     }
 
     @Override
@@ -310,117 +288,6 @@ public class ReporterServiceImpl implements ReporterService {
         return result;
     }
 
-    private NewReporter createMongoReporter(Reference reference) {
-        NewReporter newReporter = new NewReporter();
-        newReporter.setId(RandomString.generate());
-        newReporter.setEnabled(true);
-        newReporter.setName("MongoDB Reporter");
-        newReporter.setType(MONGODB);
-        newReporter.setConfiguration(createReporterConfig(reference));
-
-        return newReporter;
-    }
-
-    private NewReporter createJdbcReporter(Reference reference) {
-        NewReporter newReporter = new NewReporter();
-        newReporter.setId(RandomString.generate());
-        newReporter.setEnabled(true);
-        newReporter.setName("JDBC Reporter");
-        newReporter.setType(REPORTER_AM_JDBC);
-        newReporter.setConfiguration(createReporterConfig(reference));
-
-        return newReporter;
-    }
-
-    @Override
-    public String createReporterConfig(Reference reference) {
-        String reporterConfig = null;
-        if (useMongoReporter()) {
-            Optional<String> mongoServers = getMongoServers(environment);
-            String mongoHost = null;
-            String mongoPort = null;
-            if (mongoServers.isEmpty()) {
-                mongoHost = environment.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".mongodb.host", "localhost");
-                mongoPort = environment.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".mongodb.port", "27017");
-            }
-
-            final String username = environment.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".mongodb.username");
-            final String password = environment.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".mongodb.password");
-            String mongoDBName = getMongoDatabaseName(environment);
-
-            String defaultMongoUri = "mongodb://";
-            if (StringUtils.hasLength(username) && StringUtils.hasLength(password)) {
-                defaultMongoUri += username + ":" + password + "@";
-            }
-            defaultMongoUri += mongoServers.orElse(mongoHost + ":" + mongoPort) + "/" + mongoDBName;
-            String mongoUri = environment.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".mongodb.uri", addOptionsToURI(environment, defaultMongoUri));
-
-            var collectionSuffix = (reference == null || reference.matches(ReferenceType.ORGANIZATION, Organization.DEFAULT))
-                    ? ""
-                    : ("_" + reference.id());
-            reporterConfig = """
-                        {
-                          "uri": "%s",
-                          "host": "%s",
-                          "port": %d,
-                          "enableCredentials": false,
-                          "database": "%s",
-                          "reportableCollection": "reporter_audits%s",
-                          "bulkActions": 1000,
-                          "flushInterval": 5
-                        }
-                        """.formatted(
-                    mongoUri,
-                    (mongoHost != null) ?  mongoHost : "",
-                    (mongoPort != null) ? Integer.parseInt(mongoPort) : null,
-                    mongoDBName,
-                    collectionSuffix
-            );
-        } else if (useJdbcReporter()) {
-            String jdbcHost = environment.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".jdbc.host");
-            String jdbcPort = environment.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".jdbc.port");
-            String jdbcDatabase = environment.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".jdbc.database");
-            String jdbcDriver = environment.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".jdbc.driver");
-            String jdbcUser = environment.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".jdbc.username");
-            String jdbcPwd = environment.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".jdbc.password");
-            String jdbcSchema = environment.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".jdbc.schema");
-
-            String options = jdbcSchema == null || jdbcSchema.isEmpty() ? "[]" : """
-                    [{"option":"currentSchema","value":"%s"}]
-                    """.formatted(jdbcSchema);
-
-            reporterConfig = """
-                        {
-                          "host": "%s",
-                          "port": %d,
-                          "database": "%s",
-                          "driver": "%s",
-                          "username": "%s",
-                          "password": %s,
-                          "tableSuffix": "%s",
-                          "options": %s,
-                          "initialSize": 0,
-                          "maxSize": 10,
-                          "maxIdleTime": 30000,
-                          "maxLifeTime": 30000,
-                          "bulkActions": 1000,
-                          "flushInterval": 5
-                        }
-                        """.formatted(
-                    jdbcHost,
-                    Integer.parseInt(jdbcPort),
-                    jdbcDatabase,
-                    jdbcDriver,
-                    jdbcUser,
-                    jdbcPwd == null ? null : "\"" + jdbcPwd + "\"",
-                    getReporterTableSuffix(reference),
-                    options
-            );
-
-        }
-        return reporterConfig;
-    }
-
     @Override
     public Completable notifyInheritedReporters(Reference parentReference, Reference childReporterReference, Action childReporterAction) {
         return reporterRepository.findByReference(parentReference)
@@ -432,75 +299,5 @@ public class ReporterServiceImpl implements ReporterService {
                     return eventService.create(event);
                 })
                 .ignoreElements();
-    }
-
-    private static String getReporterTableSuffix(Reference reference) {
-        if (reference == null || reference.matches(ReferenceType.ORGANIZATION, Organization.DEFAULT)) {
-            return "";
-        }
-        // dashes are forbidden in table names, replace them in domainName by underscore
-        var tableSuffix = reference.id().replace("-", "_");
-        if (tableSuffix.length() <= TABLE_SUFFIX_MAX_LENGTH) {
-            return tableSuffix;
-        }
-        try {
-            LOGGER.info("Table name 'reporter_audits_access_points_{}' will be too long, compute shortest unique name", tableSuffix);
-            byte[] hash = MessageDigest.getInstance("sha-256").digest(tableSuffix.getBytes());
-            return BaseEncoding.base16().encode(hash).substring(0, 30).toLowerCase();
-        } catch (NoSuchAlgorithmException e) {
-            throw new IllegalStateException("Unable to compute digest of '" + reference.id() + "' due to unknown sha-256 algorithm", e);
-        }
-    }
-
-    private String addOptionsToURI(RepositoriesEnvironment environment, String mongoUri) {
-        Integer connectTimeout = environment.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".mongodb.connectTimeout", Integer.class, 5000);
-        Integer socketTimeout = environment.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".mongodb.socketTimeout", Integer.class, 5000);
-        Integer maxConnectionIdleTime = environment.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".mongodb.maxConnectionIdleTime", Integer.class);
-        Integer heartbeatFrequency = environment.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".mongodb.heartbeatFrequency", Integer.class);
-        Boolean sslEnabled = environment.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".mongodb.sslEnabled", Boolean.class);
-        String authSource = environment.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".mongodb.authSource", String.class);
-
-        mongoUri += "?connectTimeoutMS=" + connectTimeout + "&socketTimeoutMS=" + socketTimeout;
-        if (authSource != null) {
-            mongoUri += "&authSource=" + authSource;
-        }
-        if (maxConnectionIdleTime != null) {
-            mongoUri += "&maxIdleTimeMS=" + maxConnectionIdleTime;
-        }
-        if (heartbeatFrequency != null) {
-            mongoUri += "&heartbeatFrequencyMS=" + heartbeatFrequency;
-        }
-        if (sslEnabled != null) {
-            mongoUri += "&ssl=" + sslEnabled;
-        }
-
-        return mongoUri;
-    }
-
-    private Optional<String> getMongoServers(RepositoriesEnvironment env) {
-        LOGGER.debug("Looking for MongoDB server configuration...");
-        boolean found = true;
-        int idx = 0;
-        List<String> endpoints = new ArrayList<>();
-
-        while (found) {
-            String serverHost = env.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".mongodb.servers[" + (idx++) + "].host");
-            int serverPort = env.getProperty(Scope.MANAGEMENT.getRepositoryPropertyKey() + ".mongodb.servers[" + (idx++) + "].port", int.class, 27017);
-            found = (serverHost != null);
-            if (found) {
-                endpoints.add(serverHost + ":" + serverPort);
-            }
-        }
-        return endpoints.isEmpty() ? Optional.empty() : Optional.of(endpoints.stream().collect(Collectors.joining(",")));
-    }
-
-    private boolean useMongoReporter() {
-        String managementBackend = this.environment.getProperty(MANAGEMENT_TYPE, MONGODB);
-        return MONGODB.equalsIgnoreCase(managementBackend);
-    }
-
-    private boolean useJdbcReporter() {
-        String managementBackend = this.environment.getProperty(MANAGEMENT_TYPE, MONGODB);
-        return "jdbc".equalsIgnoreCase(managementBackend);
     }
 }
